Messages can be re-sent by clicking on history
A previously published message can be re-sent by clicking on it in the
history view.  There is an outstanding issue with this whereby if a
message has not been previously published this session the sending will
fail due to no connection and the app will no longer be able to send
any messages.